Add camera's LookAt Transform

+++ src/core/vector.h
@@ -190,8 +190,8 @@ T Dot(const Vector3<T> u, const Vector3<T> v) {
 
 template<typename T>
 Vector3<T> Cross(const Vector3<T> u, const Vector3<T> v) {
-	double ux = u.x, uy = u.y, uz = u.z;
-	double vx = v.x, vy = v.y, vz = v.z;
+	Float ux = u.x, uy = u.y, uz = u.z;
+	Float vx = v.x, vy = v.y, vz = v.z;
 	return Vector3<T>(
 		uy * vz - uz * vy,
 		uz * vx - ux * vz,
